Drop the idea of multiple limits per limiter
Multiple limits per limiter is arguably easier-but-not-simpler. Requires explanation, and the user might have to wonder what happens under the hood. Which limit was exceeded? Which buckets consumed a token?

If users want multiple limits, they should create multiple limiters and call them as they wish. More explicit and observable.