Composite key for buckets (#9)
Let's try a composite key for buckets

The goal is that we can ensure that the bucket we use is for a particular limit. Buckets created with one limit but evaluated against another will be wrong or at least confusing. So let's try to make that impossible.

This is in preparation for (again!) attempting multiple limits-per-limiter.