Add garbage collection on buckets (#32)
`bucketMap` will currently grow unbounded. This PR implements a GC to remove no-longer-valuable buckets.

A bucket is no longer valuable when it is full due to the passage of time. A non-existent bucket is semantically equivalent to a full bucket; we create buckets full.

- [x] Add `isFull` method to `bucket`, which tests that its time is before or equal to now - limit.period.
- [x] `gc` method on `bucketMap`. Loop through all keys (buckets), delete if `isFull`
- [x] Add public methods `Limiter[s].GC` and `Clear`

A design decision here is not to run GC on the user’s behalf. We offer the primitive, and the user can choose timing.